[style] #272 코드 분리
